Retrieve an attachment to a message

This is a low level function to retrieve an attachment to a message by id of the attachment and message. Most users are better off using gm_save_attachments() to automatically save all the attachments in a given message.

Usage

gm_attachment(id, message_id, user_id = "me")

Arguments

id

id of the attachment

message_id

id of the parent message

user_id

gmail user_id to access, special value of 'me' indicates the authenticated user.

References

https://developers.google.com/gmail/api/reference/rest/v1/users.messages.attachments/get

See also

Examples

if (FALSE) {
my_attachment = attachment('a32e324b', '12345')
# save attachment to a file
gm_save_attachment(my_attachment, 'photo.jpg')
}