These functions are deprecated and will be removed in a future release of gmailr #'

These functions are deprecated and will be removed in a future release of gmailr #'

id(x, ...)

to(x, ...)

from(x, ...)

cc(x, ...)

bcc(x, ...)

date(x, ...)

subject(x, ...)

body(x, ...)

message(...)

messages(...)

trash_message(...)

untrash_message(...)

delete_message(...)

modify_message(...)

attachment(...)

save_attachment(...)

save_attachments(...)

insert_message(...)

import_message(...)

send_message(...)

threads(...)

thread(...)

trash_thread(...)

untrash_thread(...)

delete_thread(...)

modify_thread(...)

draft(...)

drafts(...)

create_draft(...)

send_draft(...)

labels(...)

label(...)

update_label(...)

update_label_patch(...)

delete_label(...)

create_label(...)

history(...)

mime(...)

text_body(...)

html_body(...)

attach_part(...)

attach_file(...)